requirements.txt
annotated-types==0.7.0
asttokens==3.0.0
attrs==24.2.0
ccxt==4.4.38
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
cryptography==44.0.0
dateparser==1.2.0
debugpy==1.8.11
decorator==5.1.1
executing==2.1.0
frozenlist==1.5.0
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==2.2.0  
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
loguru==0.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
numpy==1.26.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycares==4.5.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.3.0 
pydantic_core==2.6.3
Pygments==2.18.0
python-binance==1.0.25
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.31.0
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
ta==0.11.0
tenacity==9.0.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
wcwidth==0.2.13
websockets==11.0.3
yarl==1.18.3
Flask
Flask-Static-Compress